% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\name{gerar_dados}
\alias{gerar_dados}
\title{Gerar dados da RAIS do ES}
\usage{
gerar_dados()
}
\value{
TRUE/FALSE dizendo se os arquivos foram criados ou não
}
\description{
Gerar dados da RAIS do ES
}
\examples{
# Essa função não deve ser rodada. É apenas a documentação
# do processo de gerar os dados do pacote.
}
